Added HUDPanel and Bullet Casing Particle System
Fixed all compatibility issues that prevented the HUD and the bullet
Casing Particle System from being properly implemented.
The HUD displays playerHP, remaining ammo, the time, and player
identification.

The Bullet Casing Particle system simulates the bullet shell as it is
ejected sideways from the gun.

Setting default weapon to pistol is disabled in this version, and is to
be fixed by Chris.