Allowed sprite swapping for multiple players
Each player now has its own SpriteManager, allowing the sprites to swap
(for different directional facing) for multiple players.
A public reference of SpriteManager was made on Aiming, and
ChangeSprite() is called via this reference.
As Player is a prefab, this can all be hooked up via prefabs and
therefore duplicated for extra players, without needing to change
anything (in theory, yet to be tested).