Added in Giorgio's code from 24/02/
Added in Giorgio's code that was provided for 24th Feb, the day he could
not attend due to illness.

Code changes are from the text document he provided, which is located
on Dropbox in CGD Group->Shellshock->Giorgio2402.

From my understanding the new code makes the helicopter drop off a new
ammo drop every 10 seconds, and spawns a new weapon on the ammo crate
when the player collides with the ammo crate.
There is however a compilation error due to the new ammo not switching
to the player correctly.