Players can now respawn after dying.
Respawning now works.
• After the player dies, their active components are disabled (so they
can’t move, shoot or emit anything). This replaces disabling the player
game object entirely, as the Player.cs script on the player game object
still needs to be accessed, and it can’t be accessed properly (booleans)
without the gameobject it is attached to being enabled.

• Then, the player’s health, gun and ammo are reset to a default setting
(max hp, switch gun to pistol, max ammo in pistol).

• The respawn timer is then started. After the timer hits the threshold,
the player is then enabled and shown.